requirements.txt
aiohttp==3.7.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
colorama==0.4.4
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.5.1
mccabe==0.6.1
multidict==5.0.0
pylint==2.6.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
yarl==1.6.2
